requirements.txt
fastapi
uvicorn
python-telegram-bot
gspread
oauth2client
requests
httpx

